echosystem/maipp/pyproject.toml
[project]
name = "maipp"
version = "0.1.0"
description = ""
authors = [
    {name = "google-labs-jules[bot]",email = "161369871+google-labs-jules[bot]@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "boto3 (>=1.38.40,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.88.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "anthropic (>=0.54.0,<0.55.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "neo4j (>=5.28.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
moto = {extras = ["kms", "s3", "sqs"], version = "^5.1.6"}
mongomock = "^4.3.0"
